refactor: Screen 구조 세분화
