+++ src/js/util/template.js
@@ -14,9 +14,9 @@ export const SKELETON_TEMPLATE = `
     src=${notFoundImage}
     alt="video-item-thumbnail" class="video-item__thumbnail">
 </div>
-<h4 class="video-item__title"></h4>
-<p class="video-item__channel-name"></p>
-<p class="video-item__published-date"></p>
+<h4 class="video-item__title">empty</h4>
+<p class="video-item__channel-name">empty</p>
+<p class="video-item__published-date">empty</p>
 <button class="video-item__save-button button">⬇ 저장</button>
 </li>
 `.repeat(VIDEO_COUNT);
